requirements.txt
streamlit>=1.24.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
cryptography>=40.0.0
tqdm>=4.65.0
tenacity>=8.2.2
tiktoken>=0.4.0
psutil>=5.9.0